# projects_manager_backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/doctors/update-project-request:
    post:
      summary: 'Update project request'
      operationId: updateProjectRequest
      description: "Accept or reject existing project request .\nThis action should done by doctor who this project request was sent to him"
      parameters: []
      responses: {  }
      tags:
        - Projects
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                project_request_id:
                  type: integer
                  description: 'ID for this project request.'
                  example: 10
                new_status:
                  type: integer
                  description: 'New status for this project request .'
                  example: 1
                  enum:
                    - 0
                    - 1
  /api/students/project-request:
    post:
      summary: 'Send project request'
      operationId: sendProjectRequest
      description: "Make new project request and send it to specific doctor .\nThis action should be done by student who is a group admin ."
      parameters: []
      responses: {  }
      tags:
        - Projects
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                doctor_id:
                  type: integer
                  description: 'The ID of doctor .'
                  example: 1
                project_title:
                  type: string
                  description: 'Title for new project .'
                  example: 'Projects Manager'
                project_about:
                  type: string
                  description: 'About the project .'
                  example: 'This is our new project'
...